Tiding ordering and added serialize_function method.
Tidied up the the order in which the code is generated in "generated.h",
so that the struct meta data and function meta data are more seperate.
Also, moved the forward declares to the bottom.

I also added a simple utility funciton, which will serialize a function
out output. Right now, it will just throw an error for an overloaded
function.