Added a C version.
I now made another test folder, which tests the project as a .c file,
rather than .cpp. Right now, the project just reads the extension of the
first file passed in, and if that's a .c file, the generated code is a C
file. If it's a .cpp file, the generated code is a C++ file.

NOTE: I don't _really_ care about compatibility with C, but since the
generated code doesn't really use any CPP features, I figured, "why
not?"